refactor(cli): extract retro command family into cli/commands/retro.ts
